清理 clang-tidy 告警
- read_script：用带错误检查的 fseek 替代 rewind，增强文件读取鲁棒性。
- resolve：拆分 IPv6 / IPv4 分支为独立 if，控制流更清晰。
- kmod/tutu：对 csum_partial 的长度参数显式转为 int，匹配内核 API，修复潜在窄化问题。
- ktuctl/tuctl：理顺 write_to_sysfs 的 size_t/ssize_t 使用，并简化 get_boot_seconds 的错误分支。
- tucrypto：去掉无意义的 uint32_t 强制转换，保留原始无符号溢出语义。
- ratelimiter：统一 64 位常量后缀为 ULL，简化 key_equal 与令牌桶更新的写法，逻辑保持不变。

+++ kmod/tutu.c
@@ -720,7 +720,7 @@ static int skb_change_type(struct sk_buff *skb, u32 ip_type, u32 l2_len, u32 ip_
       iph             = ip_hdr(skb);
       icmph           = (typeof(icmph)) (skb->data + l4_offset);
       icmph->checksum = 0;
-      icmph->checksum = csum_fold(csum_partial((char *) icmph, icmp_len, 0));
+      icmph->checksum = csum_fold(csum_partial((char *) icmph, (int) icmp_len, 0));
       skb->ip_summed  = CHECKSUM_UNNECESSARY;
     } else if (skb->ip_summed == CHECKSUM_PARTIAL) {
       // skb->csum_start: 不变，因为udp->icmp并没有修改包长度
@@ -746,7 +746,7 @@ static int skb_change_type(struct sk_buff *skb, u32 ip_type, u32 l2_len, u32 ip_
       // 同样为了兼容性修复icmpv6检验和
       icmp6h->icmp6_cksum = 0;
       // 计算 ICMPv6 校验和(带伪头部)
-      __wsum csum         = csum_partial((char *) icmp6h, icmp_len, 0);
+      __wsum csum         = csum_partial((char *) icmp6h, (int) icmp_len, 0);
       icmp6h->icmp6_cksum = csum_ipv6_magic(&ip6h->saddr, &ip6h->daddr, icmp_len, IPPROTO_ICMPV6, csum);
       skb->ip_summed      = CHECKSUM_UNNECESSARY;
     } else if (skb->ip_summed == CHECKSUM_PARTIAL) {
